# MVP/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: World of Shadows — Backend API
  version: 1.0.0
  description: >-
    Bounded machine-readable API schema for the Flask backend /api/v1 HTTP surface.
    This artifact is authoritative for route inventory and OpenAPI-tag grouping inside
    this MVP package, but it is not the sole upstream authority for all overlapping
    runtime semantics. The higher-order runtime/API authority for play-service run
    lifecycle envelopes remains docs/technical/architecture/canonical_runtime_contract.md,
    and world-engine runtime authority remains governed by docs/technical/runtime/runtime-authority-and-state-flow.md
    and docs/ADR/adr-0001-runtime-authority-in-world-engine.md.
  x-wos-governance:
    artifact_role: canonical_api_schema
    authority_scope: bounded_backend_api_v1_route_inventory
    upstream_contracts:
      - docs/technical/architecture/canonical_runtime_contract.md
      - docs/technical/runtime/runtime-authority-and-state-flow.md
      - docs/ADR/adr-0001-runtime-authority-in-world-engine.md
    not_authoritative_for:
      - world-engine fastapi surface outside documented backend projection
      - websocket message contracts
      - runtime committed truth
      - unpublished or operator-only convenience notes
servers:
- url: /
  description: Same origin as backend (set base URL in your client).
...
paths:
...
  /api/v1/data/export:
    post:
      tags:
      - Admin
      summary: POST /data/export
      operationId: post_api_v1_data_export
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
  /api/v1/data/export/decrypt:
    post:
      tags:
      - Admin
      summary: POST /data/export/decrypt
      operationId: post_api_v1_data_export_decrypt
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
  /api/v1/data/import/execute:
    post:
      tags:
      - Admin
      summary: POST /data/import/execute
      operationId: post_api_v1_data_import_execute
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
  /api/v1/data/import/preflight:
    post:
      tags:
      - Admin
      summary: POST /data/import/preflight
      operationId: post_api_v1_data_import_preflight
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Flask-JWT-Extended access token from `POST /api/v1/auth/login`.
security:
- BearerAuth: []